requirements.txt
streamlit
pandas
Pillow
ultralytics
torch
numpy
